properly display details of system status items
the help text element in the status window of stus-ui.xml
does contain multiline details text for the selected status item.
the category list and the log list are now on top of each other
and either list hides when the user selects the other list. This
makes it consistent with other screen containing multiple button
lists like the recordings screen.