Show watched/recorded percent progress
This commit adds watched and recorded percent progress
to the recording list via a progress bar as as well as
to the recording details as text and corresponding icon
of the "watchrecordings" screen.
The recording progress as progress bar is also added to
the scheduled list of the "viewscheduled" screen.